requirements.txt
langchain
langchain-huggingface
langchain-groq
langchain-community
beautifulsoup4
python-dotenv
faiss-cpu
sentence-transformers
huggingface-hub
streamlit